feat(auth): implement authentication page logic
- Create a custom hook `useAuthPageLogic` for handling navigation.
- Refactor `AuthScreen` to utilize the new hook for account creation, login, and guest navigation.
- Remove redundant background color setting from `AuthScreen`.
